Move contract code into it's own directory
The contract code is becoming a bit too complex for a single source file.
