+++ src/main.cpp
@@ -22,7 +22,6 @@
 #include "neuralnet/tally.h"
 #include "backup.h"
 #include "appcache.h"
-#include "contract/message.h"
 #include "scraper_net.h"
 #include "gridcoin.h"
