+++ src/neuralnet/contract/contract.cpp
@@ -1,6 +1,6 @@
 #include "appcache.h"
 #include "main.h"
-#include "neuralnet/contract.h"
+#include "neuralnet/contract/contract.h"
 #include "neuralnet/beacon.h"
 #include "neuralnet/project.h"
 #include "util.h"
